Move state to SingleArticle component and move util function out
