+++ app/javascript/modCenter/moderationArticles.jsx
@@ -5,7 +5,6 @@ import SingleArticle from './singleArticle';
 export class ModerationArticles extends Component {
   state = {
     articles: [],
-    articleOpened: false,
   };
 
   componentWillMount() {
@@ -15,21 +14,6 @@ export class ModerationArticles extends Component {
       articles,
     });
   }
-
-  toggleArticle = (e, id, path) => {
-    e.preventDefault();
-
-    const { articleOpened } = this.state;
-    if (articleOpened) {
-      this.setState({ articleOpened: false });
-      document.getElementById(`article-iframe-${id}`).innerHTML = '';
-    } else {
-      this.setState({ articleOpened: true });
-      document.getElementById(`article-iframe-${id}`).innerHTML = `
-  <iframe class="article-iframe" src="${path}"></iframe><iframe class="actions-panel-iframe" src="${path}/actions_panel"></iframe>
-      `;
-    }
-  };
 
   render() {
     const { articles } = this.state;
@@ -53,7 +37,6 @@ export class ModerationArticles extends Component {
               cachedTagList={cachedTagList}
               publishedAt={publishedAt}
               user={user}
-              toggleArticle={this.toggleArticle}
             />
           );
         })}
